ResqueListener split into Loner and Throttle
