+++ Loner.php
@@ -4,13 +4,10 @@ namespace ShonM\ResqueBundle\Listener;
 
 use Doctrine\Common\Annotations\Reader;
 use ShonM\ResqueBundle\Resque;
-use ShonM\ResqueBundle\Annotation\Throttled;
 use ShonM\ResqueBundle\Annotation\Loner;
-use ShonM\ResqueBundle\Exception\ThrottledException;
-use ReflectionClass;
 use Resque\Job;
 
-class ResqueListener
+class Loner
 {
     protected $resque;
 
@@ -20,49 +17,6 @@ class ResqueListener
     {
         $this->resque = $resque;
         $this->annotationReader = $annotationReader;
-    }
-
-    // Not referenced
-    public function onBeforeEnqueue($eventArg)
-    {
-        $class = $eventArg->getClass();
-        $throttle = $this->getThrottleAnnotation($class);
-        if ($throttle) {
-            $throttleKey = $this->throttleKey($class, $throttle);
-            if ($this->shouldThrottle($throttleKey, $throttle)) {
-                throw new ThrottledException("'$class' with key '$throttleKey' has exceeded it's throttle limit.");
-            }
-        }
-    }
-
-    protected function getThrottleAnnotation($class)
-    {
-        return $this->annotationReader->getClassAnnotation(
-            new ReflectionClass($class),
-            'ShonM\ResqueBundle\Annotation\Throttled'
-        );
-    }
-
-    protected function shouldThrottle($key, Throttled $throttle)
-    {
-        $redis = $this->resque->redis();
-        $result = (bool) $redis->exists($key);
-        if ($result === false) {
-            $redis->set($key, true, $throttle->canRunEvery);
-        }
-
-        return $result;
-    }
-
-    protected function throttleKey($class, $throttle)
-    {
-        $key = $class;
-        if ($throttle->keyMethod) {
-            $method = $throttle->keyMethod;
-            $key = $class::$method($throttle);
-        }
-
-        return "throttle:$key";
     }
 
     protected function lonerKey($queue, $item, Loner $loner)
